FIX TrainingArena scene incorrectly serializing in binary format
Apparently NavMesh data is in binary by design. The way TilemapNavMeshGenerator was updating NavMesh data, was storing the NavMesh directly in the scene, instead of saving it in a saparat asset. Having NavMesh data embeded in the scene forced the scene to be saved in binary format.

We have refactored TilemapNavMeshGenerator to save the generated NavMesh in a separate asset instead of the scene.